Put css formatter into it's own file
